Fix session restore of Reader-Mode and Error Pages
