+++ ios/brave-ios/App/iOS/Delegates/AppState.swift
@@ -205,7 +205,6 @@ public class AppState {
     let responders: [(String, InternalSchemeResponse)] = [
       (AboutHomeHandler.path, AboutHomeHandler()),
       (AboutLicenseHandler.path, AboutLicenseHandler()),
-      (SessionRestoreHandler.path, SessionRestoreHandler()),
       (ErrorPageHandler.path, ErrorPageHandler()),
       (ReaderModeHandler.path, ReaderModeHandler(profile: profile)),
       (Web3DomainHandler.path, Web3DomainHandler()),
