+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+TabWebPolicyDecider.swift
@@ -117,10 +117,6 @@ extension BrowserViewController: TabWebPolicyDecider {
       let internalUrl = InternalURL(responseURL)
 
       tab.rewardsReportingState.httpStatusCode = response.statusCode
-
-      if !tab.rewardsReportingState.wasRestored {
-        tab.rewardsReportingState.wasRestored = internalUrl?.isSessionRestore == true
-      }
     }
 
     let request = response.url.flatMap { pendingRequests[$0.absoluteString] }
@@ -209,11 +205,7 @@ extension BrowserViewController: TabWebPolicyDecider {
     }
 
     // Handle internal:// urls
-    if InternalURL.isValid(url: requestURL) {
-      if requestInfo.navigationType != .backForward, request.isInternalUnprivileged {
-        Logger.module.warning("Denying unprivileged request: \(request)")
-        return .cancel
-      }
+    if InternalURL.isValid(url: requestURL), requestInfo.navigationType != .backForward {
       return .allow
     }
 
@@ -552,7 +544,6 @@ extension BrowserViewController: TabWebPolicyDecider {
       // request then the page is reloaded with a proper url and adblocking rules are applied.
       if let mainDocumentURL = request.mainDocumentURL,
         mainDocumentURL.schemelessAbsoluteString == requestURL.schemelessAbsoluteString,
-        !(InternalURL(requestURL)?.isSessionRestore ?? false),
         requestInfo.isMainFrame
       {
         // Identify specific block lists that need to be applied to the requesting domain
