+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Interstitial Pages/InternalSchemeHandler.swift
@@ -177,3 +177,19 @@ public class InternalSchemeHandler: NSObject, WKURLSchemeHandler {
     activeTasks.removeObject(forKey: urlSchemeTask)
   }
 }
+
+extension WKWebView {
+  // Use JS to redirect the page without adding a history entry
+  func replaceLocation(with url: URL) {
+    let apostropheEncoded = "%27"
+    let safeUrl = url.absoluteString.replacingOccurrences(of: "'", with: apostropheEncoded)
+    evaluateSafeJavaScript(
+      functionName: "location.replace",
+      args: ["'\(safeUrl)'"],
+      contentWorld: .defaultClient,
+      escapeArgs: false,
+      asFunction: true,
+      completion: nil
+    )
+  }
+}
