+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Playlist/Managers & Cache/PlaylistCacheLoader.swift
@@ -392,7 +392,6 @@ extension LivePlaylistWebLoader: WKNavigationDelegate {
 
       if let mainDocumentURL = navigationAction.request.mainDocumentURL,
         mainDocumentURL.schemelessAbsoluteString == url.schemelessAbsoluteString,
-        !(InternalURL(url)?.isSessionRestore ?? false),
         navigationAction.sourceFrame.isMainFrame
           || navigationAction.targetFrame?.isMainFrame == true
       {
