+++ ios/brave-ios/Sources/BraveShared/Extensions/URLExtensions.swift
@@ -32,7 +32,7 @@ extension URL {
       switch internalURL.urlType {
       case .errorPage:
         return internalURL.originalURLFromErrorPage
-      case .web3Page, .sessionRestorePage, .aboutHomePage:
+      case .web3Page, .aboutHomePage:
         return internalURL.extractedUrlParam
       case .blockedPage:
         return decodeEmbeddedInternalURL(for: .blocked)
@@ -68,7 +68,7 @@ extension URL {
     }
 
     if let internalUrl = InternalURL(self),
-      internalUrl.isSessionRestore || internalUrl.isWeb3URL || internalUrl.isHTTPBlockedPage
+      internalUrl.isWeb3URL || internalUrl.isHTTPBlockedPage
         || internalUrl.isBlockedPage
     {
       return internalUrl.extractedUrlParam?.displayURL
@@ -280,7 +280,6 @@ extension InternalURL {
   enum URLType {
     case blockedPage
     case httpBlockedPage
-    case sessionRestorePage
     case errorPage
     case readerModePage
     case aboutHomePage
@@ -310,10 +309,6 @@ extension InternalURL {
 
     if isReaderModePage {
       return .readerModePage
-    }
-
-    if isSessionRestore {
-      return .sessionRestorePage
     }
 
     if isAboutHomeURL {
