+++ ios/brave-ios/Tests/ClientTests/ClientTests.swift
@@ -21,7 +21,6 @@ class ClientTests: XCTestCase {
     let responders: [(String, InternalSchemeResponse)] = [
       (AboutHomeHandler.path, AboutHomeHandler()),
       (AboutLicenseHandler.path, AboutLicenseHandler()),
-      (SessionRestoreHandler.path, SessionRestoreHandler()),
       (ErrorPageHandler.path, ErrorPageHandler()),
       (ReaderModeHandler.path, ReaderModeHandler(profile: BrowserProfile(localName: "profile"))),
     ]
